refactor(chat-input): organize imports
