+++ apps/web/src/components/features/ChatInput/Emoji/EmojiSuggestion.tsx
