+++ apps/web/src/components/features/ChatInput/Mentions/createMentionSuggestion.tsx
